refactor: modularize settings component and converge assembly layer (AUDIT-002 Phase B)
- ARCH-002: split SettingsComponent into declaration header and SettingsComponent.cpp implementation, breaking down buildJiveUi into modular audio/midi/appearance wiring routines.
- ARCH-001: extract inline binding edit merge and preset saving logic from MainComponent::initialiseUi into dedicated handlers and PresetFlowSupport::autoSaveCurrentPreset.
- QUAL-002: create shared JiveBuilderHelpers.h and remove duplicate node/text/button/flexRow helpers across 4 layout files.
- QUAL-007: remove duplicated JIVE teardown helpers from MainComponent.cpp and reuse standard JiveUtils.h functions.
- Update AUDIT-002 review record 2 and problem table to close ARCH-001, ARCH-002, QUAL-002, and QUAL-007.
