+++ source/UI/KeyBindingEditDialog.cpp
@@ -7,67 +7,13 @@
 
 #include "UI/ColourSwatchButton.h"
 #include "UI/jive/DesignTokens.h"
+#include "UI/jive/JiveBuilderHelpers.h"
 #include "UI/jive/JiveModalDialog.h"
 #include "UI/jive/JiveUtils.h"
 #include "UI/jive/StyleCatalog.h"
 
 namespace {
 using namespace devpiano::ui::jive;
-inline juce::ValueTree node(const juce::Identifier& type, const juce::String& id = {}) {
-    auto t = juce::ValueTree(type);
-    if (id.isNotEmpty()) {
-        t.setProperty("id", id, nullptr);
-    }
-    return t;
-}
-
-inline juce::ValueTree text(const juce::String& content, const juce::String& id = {}) {
-    auto t = node("Text", id);
-    t.setProperty("text", content, nullptr);
-    t.setProperty("title", content, nullptr);
-    return t;
-}
-
-inline juce::ValueTree button(const juce::String& label, const juce::String& id = {}) {
-    auto t = node("Button", id);
-    t.setProperty("display", "flex", nullptr);
-    t.setProperty("justify-content", "centre", nullptr);
-    t.setProperty("align-items", "centre", nullptr);
-    t.setProperty("title", label, nullptr);
-    t.setProperty("border-width", "1", nullptr);
-
-    auto labelText = text(label, id.isNotEmpty() ? id + "-text" : juce::String {});
-    labelText.setProperty("justification", "centred", nullptr);
-    labelText.setProperty("word-wrap", "none", nullptr);
-    t.appendChild(labelText, nullptr);
-
-    return t;
-}
-
-inline juce::ValueTree flexRow(const juce::String& id = {}) {
-    auto t = node("Component", id);
-    t.setProperty("display", "flex", nullptr);
-    t.setProperty("flex-direction", "row", nullptr);
-    t.setProperty("align-items", "centre", nullptr);
-    return t;
-}
-
-inline juce::ValueTree settingRow(const juce::String& labelStr, const juce::ValueTree& controlNode,
-                                  const juce::String& labelId = {}) {
-    auto row = flexRow();
-    row.setProperty("height", 28, nullptr);
-    row.setProperty("margin", "0 0 6 0", nullptr);
-
-    auto lbl = text(labelStr, labelId);
-    lbl.setProperty("flex-grow", 1.0, nullptr);
-    lbl.setProperty("height", 22, nullptr);
-    lbl.setProperty("font-size", 14, nullptr);
-    lbl.setProperty("justification", "centred-left", nullptr);
-    row.appendChild(lbl, nullptr);
-
-    row.appendChild(controlNode, nullptr);
-    return row;
-}
 
 const std::array<juce::Colour, 8> paletteColours {
     juce::Colour(0xFF00C8FF), // Cyan
